+++ include/api.php
@@ -1015,127 +1015,6 @@ function api_media_metadata_create($type)
 api_register_func('api/media/metadata/create', 'api_media_metadata_create', true, API_METHOD_POST);
 
 /**
- * Returns extended information of a given user, specified by ID or screen name as per the required id parameter.
- * The author's most recent status will be returned inline.
- *
- * @param string $type Return type (atom, rss, xml, json)
- * @return array|string
- * @throws BadRequestException
- * @throws ImagickException
- * @throws InternalServerErrorException
- * @throws UnauthorizedException
- * @see https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-show
- */
-function api_users_show($type)
-{
-	BaseApi::checkAllowedScope(BaseApi::SCOPE_READ);
-	$uid = BaseApi::getCurrentUserID();
-
-	$user_info = DI::twitterUser()->createFromUserId($uid, false)->toArray();
-
-	// "uid" is only needed for some internal stuff, so remove it from here
-	unset($user_info['uid']);
-
-	return DI::apiResponse()->formatData('user', $type, ['user' => $user_info]);
-}
-
-api_register_func('api/users/show', 'api_users_show');
-api_register_func('api/externalprofile/show', 'api_users_show');
-
-/**
- * Search a public user account.
- *
- * @param string $type Return type (atom, rss, xml, json)
- *
- * @return array|string
- * @throws BadRequestException
- * @throws ImagickException
- * @throws InternalServerErrorException
- * @throws UnauthorizedException
- * @see https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-search
- */
-function api_users_search($type)
-{
-	BaseApi::checkAllowedScope(BaseApi::SCOPE_READ);
-	$uid = BaseApi::getCurrentUserID();
-
-	$userlist = [];
-
-	if (!empty($_GET['q'])) {
-		$contacts = Contact::selectToArray(
-			['id'],
-			[
-				'`uid` = 0 AND (`name` = ? OR `nick` = ? OR `url` = ? OR `addr` = ?)',
-				$_GET['q'],
-				$_GET['q'],
-				$_GET['q'],
-				$_GET['q'],
-			]
-		);
-
-		if (DBA::isResult($contacts)) {
-			$k = 0;
-			foreach ($contacts as $contact) {
-				$user_info = DI::twitterUser()->createFromContactId($contact['id'], $uid, false)->toArray();
-
-				if ($type == 'xml') {
-					$userlist[$k++ . ':user'] = $user_info;
-				} else {
-					$userlist[] = $user_info;
-				}
-			}
-			$userlist = ['users' => $userlist];
-		} else {
-			throw new NotFoundException('User ' . $_GET['q'] . ' not found.');
-		}
-	} else {
-		throw new BadRequestException('No search term specified.');
-	}
-
-	return DI::apiResponse()->formatData('users', $type, $userlist);
-}
-
-api_register_func('api/users/search', 'api_users_search');
-
-/**
- * Return user objects
- *
- * @see https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup
- *
- * @param string $type Return format: json or xml
- *
- * @return array|string
- * @throws BadRequestException
- * @throws ImagickException
- * @throws InternalServerErrorException
- * @throws NotFoundException if the results are empty.
- * @throws UnauthorizedException
- */
-function api_users_lookup($type)
-{
-	BaseApi::checkAllowedScope(BaseApi::SCOPE_READ);
-	$uid = BaseApi::getCurrentUserID();
-
-	$users = [];
-
-	if (!empty($_REQUEST['user_id'])) {
-		foreach (explode(',', $_REQUEST['user_id']) as $cid) {
-			if (!empty($cid) && is_numeric($cid)) {
-				$users[] = DI::twitterUser()->createFromContactId((int)$cid, $uid, false)->toArray();
-			}
-		}
-	}
-
-	if (empty($users)) {
-		throw new NotFoundException;
-	}
-
-	return DI::apiResponse()->formatData("users", $type, ['users' => $users]);
-}
-
-api_register_func('api/users/lookup', 'api_users_lookup', true);
-
-/**
  * Returns statuses that match a specified query.
  *
  * @see https://developer.twitter.com/en/docs/tweets/search/api-reference/get-search-tweets
@@ -1242,167 +1121,6 @@ api_register_func('api/search/tweets', 'api_search', true);
 api_register_func('api/search', 'api_search', true);
 
 /**
- * Returns a single status.
- *
- * @param string $type Return type (atom, rss, xml, json)
- *
- * @return array|string
- * @throws BadRequestException
- * @throws ForbiddenException
- * @throws ImagickException
- * @throws InternalServerErrorException
- * @throws UnauthorizedException
- * @see https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-show-id
- */
-function api_statuses_show($type)
-{
-	BaseApi::checkAllowedScope(BaseApi::SCOPE_READ);
-	$uid = BaseApi::getCurrentUserID();
-
-	// params
-	$id = intval(DI::args()->getArgv()[3] ?? 0);
-
-	if ($id == 0) {
-		$id = intval($_REQUEST['id'] ?? 0);
-	}
-
-	// Hotot workaround
-	if ($id == 0) {
-		$id = intval(DI::args()->getArgv()[4] ?? 0);
-	}
-
-	logger::notice('API: api_statuses_show: ' . $id);
-
-	$conversation = !empty($_REQUEST['conversation']);
-
-	// try to fetch the item for the local user - or the public item, if there is no local one
-	$uri_item = Post::selectFirst(['uri-id'], ['id' => $id]);
-	if (!DBA::isResult($uri_item)) {
-		throw new BadRequestException(sprintf("There is no status with the id %d", $id));
-	}
-
-	$item = Post::selectFirst(['id'], ['uri-id' => $uri_item['uri-id'], 'uid' => [0, $uid]], ['order' => ['uid' => true]]);
-	if (!DBA::isResult($item)) {
-		throw new BadRequestException(sprintf("There is no status with the uri-id %d for the given user.", $uri_item['uri-id']));
-	}
-
-	$id = $item['id'];
-
-	if ($conversation) {
-		$condition = ['parent' => $id, 'gravity' => [GRAVITY_PARENT, GRAVITY_COMMENT]];
-		$params = ['order' => ['id' => true]];
-	} else {
-		$condition = ['id' => $id, 'gravity' => [GRAVITY_PARENT, GRAVITY_COMMENT]];
-		$params = [];
-	}
-
-	$statuses = Post::selectForUser($uid, [], $condition, $params);
-
-	/// @TODO How about copying this to above methods which don't check $r ?
-	if (!DBA::isResult($statuses)) {
-		throw new BadRequestException(sprintf("There is no status or conversation with the id %d.", $id));
-	}
-
-	$include_entities = strtolower(($_REQUEST['include_entities'] ?? 'false') == 'true');
-
-	$ret = [];
-	while ($status = DBA::fetch($statuses)) {
-		$ret[] = DI::twitterStatus()->createFromUriId($status['uri-id'], $status['uid'], $include_entities)->toArray();
-	}
-	DBA::close($statuses);
-
-	if ($conversation) {
-		$data = ['status' => $ret];
-		return DI::apiResponse()->formatData("statuses", $type, $data);
-	} else {
-		$data = ['status' => $ret[0]];
-		return DI::apiResponse()->formatData("status", $type, $data);
-	}
-}
-
-api_register_func('api/statuses/show', 'api_statuses_show', true);
-
-/**
- *
- * @param string $type Return type (atom, rss, xml, json)
- *
- * @return array|string
- * @throws BadRequestException
- * @throws ForbiddenException
- * @throws ImagickException
- * @throws InternalServerErrorException
- * @throws UnauthorizedException
- * @todo nothing to say?
- */
-function api_conversation_show($type)
-{
-	BaseApi::checkAllowedScope(BaseApi::SCOPE_READ);
-	$uid = BaseApi::getCurrentUserID();
-
-	// params
-	$id       = intval(DI::args()->getArgv()[3]           ?? 0);
-	$since_id = intval($_REQUEST['since_id'] ?? 0);
-	$max_id   = intval($_REQUEST['max_id']   ?? 0);
-	$count    = intval($_REQUEST['count']    ?? 20);
-	$page     = intval($_REQUEST['page']     ?? 1);
-
-	$start = max(0, ($page - 1) * $count);
-
-	if ($id == 0) {
-		$id = intval($_REQUEST['id'] ?? 0);
-	}
-
-	// Hotot workaround
-	if ($id == 0) {
-		$id = intval(DI::args()->getArgv()[4] ?? 0);
-	}
-
-	Logger::info(API_LOG_PREFIX . '{subaction}', ['module' => 'api', 'action' => 'conversation', 'subaction' => 'show', 'id' => $id]);
-
-	// try to fetch the item for the local user - or the public item, if there is no local one
-	$item = Post::selectFirst(['parent-uri-id'], ['id' => $id]);
-	if (!DBA::isResult($item)) {
-		throw new BadRequestException("There is no status with the id $id.");
-	}
-
-	$parent = Post::selectFirst(['id'], ['uri-id' => $item['parent-uri-id'], 'uid' => [0, $uid]], ['order' => ['uid' => true]]);
-	if (!DBA::isResult($parent)) {
-		throw new BadRequestException("There is no status with this id.");
-	}
-
-	$id = $parent['id'];
-
-	$condition = ["`parent` = ? AND `uid` IN (0, ?) AND `gravity` IN (?, ?) AND `id` > ?",
-		$id, $uid, GRAVITY_PARENT, GRAVITY_COMMENT, $since_id];
-
-	if ($max_id > 0) {
-		$condition[0] .= " AND `id` <= ?";
-		$condition[] = $max_id;
-	}
-
-	$params = ['order' => ['id' => true], 'limit' => [$start, $count]];
-	$statuses = Post::selectForUser($uid, [], $condition, $params);
-
-	if (!DBA::isResult($statuses)) {
-		throw new BadRequestException("There is no status with id $id.");
-	}
-
-	$include_entities = strtolower(($_REQUEST['include_entities'] ?? 'false') == 'true');
-
-	$ret = [];
-	while ($status = DBA::fetch($statuses)) {
-		$ret[] = DI::twitterStatus()->createFromUriId($status['uri-id'], $status['uid'], $include_entities)->toArray();
-	}
-	DBA::close($statuses);
-
-	$data = ['status' => $ret];
-	return DI::apiResponse()->formatData("statuses", $type, $data);
-}
-
-api_register_func('api/conversation/show', 'api_conversation_show', true);
-api_register_func('api/statusnet/conversation', 'api_conversation_show', true);
-
-/**
  * Repeats a status.
  *
  * @param string $type Return type (atom, rss, xml, json)
@@ -1480,47 +1198,6 @@ function api_statuses_repeat($type)
 }
 
 api_register_func('api/statuses/retweet', 'api_statuses_repeat', true, API_METHOD_POST);
-
-/**
- * Destroys a specific status.
- *
- * @param string $type Return type (atom, rss, xml, json)
- *
- * @return array|string
- * @throws BadRequestException
- * @throws ForbiddenException
- * @throws ImagickException
- * @throws InternalServerErrorException
- * @throws UnauthorizedException
- * @see https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/post-statuses-destroy-id
- */
-function api_statuses_destroy($type)
-{
-	BaseApi::checkAllowedScope(BaseApi::SCOPE_WRITE);
-	$uid = BaseApi::getCurrentUserID();
-
-	// params
-	$id = intval(DI::args()->getArgv()[3] ?? 0);
-
-	if ($id == 0) {
-		$id = intval($_REQUEST['id'] ?? 0);
-	}
-
-	// Hotot workaround
-	if ($id == 0) {
-		$id = intval(DI::args()->getArgv()[4] ?? 0);
-	}
-
-	logger::notice('API: api_statuses_destroy: ' . $id);
-
-	$ret = api_statuses_show($type);
-
-	Item::deleteForUser(['id' => $id], $uid);
-
-	return $ret;
-}
-
-api_register_func('api/statuses/destroy', 'api_statuses_destroy', true, API_METHOD_DELETE);
 
 /**
  * Star/unstar an item.
@@ -2551,12 +2228,12 @@ function api_account_update_profile_image($type)
 	// output for client
 	if ($data) {
 		$skip_status = $_REQUEST['skip_status'] ?? false;
-	
+
 		$user_info = DI::twitterUser()->createFromUserId($uid, $skip_status)->toArray();
-	
+
 		// "verified" isn't used here in the standard
 		unset($user_info["verified"]);
-	
+
 		// "uid" is only needed for some internal stuff, so remove it from here
 		unset($user_info['uid']);
