+++ static/routes.config.php
@@ -48,27 +48,27 @@ $apiRoutes = [
 		'/update_profile_image[.{extension:json|xml|rss|atom}]'    => [Module\Api\Friendica\Index::class,                   [        R::POST]],
 	],
 
-	'/blocks/list[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Friendica\Index::class,         [R::GET         ]],
-	'/conversation/show[.{extension:json|xml|rss|atom}]'           => [Module\Api\Friendica\Index::class,         [R::GET         ]],
+	'/blocks/list[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
+	'/conversation/show[.{extension:json|xml|rss|atom}]'           => [Module\Api\GNUSocial\Statusnet\Conversation::class, [R::GET         ]],
 	'/direct_messages' => [
-		'/all[.{extension:json|xml|rss|atom}]'                     => [Module\Api\Friendica\Index::class,         [R::GET         ]],
-		'/conversation[.{extension:json|xml|rss|atom}]'            => [Module\Api\Friendica\Index::class,         [R::GET         ]],
-		'/destroy[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Friendica\Index::class,         [R::DELETE, R::POST]],
-		'/new[.{extension:json|xml|rss|atom}]'                     => [Module\Api\Friendica\Index::class,         [        R::POST]],
-		'/sent[.{extension:json|xml|rss|atom}]'                    => [Module\Api\Friendica\Index::class,         [R::GET         ]],
+		'/all[.{extension:json|xml|rss|atom}]'                     => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
+		'/conversation[.{extension:json|xml|rss|atom}]'            => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
+		'/destroy[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Friendica\Index::class,                  [R::DELETE, R::POST]],
+		'/new[.{extension:json|xml|rss|atom}]'                     => [Module\Api\Friendica\Index::class,                  [        R::POST]],
+		'/sent[.{extension:json|xml|rss|atom}]'                    => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
 	],
-	'/direct_messages[.{extension:json|xml|rss|atom}]'             => [Module\Api\Friendica\Index::class,         [R::GET, R::POST]],
+	'/direct_messages[.{extension:json|xml|rss|atom}]'             => [Module\Api\Friendica\Index::class,                  [R::GET, R::POST]],
 
-	'/externalprofile/show[.{extension:json|xml|rss|atom}]'        => [Module\Api\Friendica\Index::class,         [R::GET         ]],
-	'/favorites/create[.{extension:json|xml|rss|atom}]'            => [Module\Api\Friendica\Index::class,         [        R::POST]],
-	'/favorites/destroy[.{extension:json|xml|rss|atom}]'           => [Module\Api\Friendica\Index::class,         [R::DELETE, R::POST]],
-	'/favorites[.{extension:json|xml|rss|atom}]'                   => [Module\Api\Twitter\Favorites::class,       [R::GET         ]],
-	'/followers/ids[.{extension:json|xml|rss|atom}]'               => [Module\Api\Twitter\Followers\Ids::class,   [R::GET         ]],
-	'/followers/list[.{extension:json|xml|rss|atom}]'              => [Module\Api\Twitter\Followers\Lists::class, [R::GET         ]],
-	'/friends/ids[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Twitter\Friends\Ids::class,     [R::GET         ]],
-	'/friends/list[.{extension:json|xml|rss|atom}]'                => [Module\Api\Twitter\Friends\Lists::class,   [R::GET         ]],
-	'/friendships/destroy[.{extension:json|xml|rss|atom}]'         => [Module\Api\Friendica\Index::class,         [        R::POST]],
-	'/friendships/incoming[.{extension:json|xml|rss|atom}]'        => [Module\Api\Friendica\Index::class,         [R::GET         ]],
+	'/externalprofile/show[.{extension:json|xml|rss|atom}]'        => [Module\Api\Twitter\Users\Show::class,               [R::GET         ]],
+	'/favorites/create[.{extension:json|xml|rss|atom}]'            => [Module\Api\Friendica\Index::class,                  [        R::POST]],
+	'/favorites/destroy[.{extension:json|xml|rss|atom}]'           => [Module\Api\Friendica\Index::class,                  [R::DELETE, R::POST]],
+	'/favorites[.{extension:json|xml|rss|atom}]'                   => [Module\Api\Twitter\Favorites::class,                [R::GET         ]],
+	'/followers/ids[.{extension:json|xml|rss|atom}]'               => [Module\Api\Twitter\Followers\Ids::class,            [R::GET         ]],
+	'/followers/list[.{extension:json|xml|rss|atom}]'              => [Module\Api\Twitter\Followers\Lists::class,          [R::GET         ]],
+	'/friends/ids[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Twitter\Friends\Ids::class,              [R::GET         ]],
+	'/friends/list[.{extension:json|xml|rss|atom}]'                => [Module\Api\Twitter\Friends\Lists::class,            [R::GET         ]],
+	'/friendships/destroy[.{extension:json|xml|rss|atom}]'         => [Module\Api\Friendica\Index::class,                  [        R::POST]],
+	'/friendships/incoming[.{extension:json|xml|rss|atom}]'        => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
 
 	'/friendica' => [
 		'/activity/{verb:attendmaybe|attendno|attendyes|dislike|like|unattendmaybe|unattendno|unattendyes|undislike|unlike}[.{extension:json|xml|rss|atom}]'
@@ -106,15 +106,15 @@ $apiRoutes = [
 		'/update[.{extension:json|xml|rss|atom}]'                  => [Module\Api\Friendica\Index::class, [        R::POST]],
 	],
 
-	'/media/upload[.{extension:json|xml|rss|atom}]'                    => [Module\Api\Friendica\Index::class,             [        R::POST]],
-	'/media/metadata/create[.{extension:json|xml|rss|atom}]'           => [Module\Api\Friendica\Index::class,             [        R::POST]],
-	'/saved_searches/list[.{extension:json|xml|rss|atom}]'             => [Module\Api\Twitter\SavedSearches::class,       [R::GET         ]],
-	'/search/tweets[.{extension:json|xml|rss|atom}]'                   => [Module\Api\Friendica\Index::class,             [R::GET         ]],
-	'/search[.{extension:json|xml|rss|atom}]'                          => [Module\Api\Friendica\Index::class,             [R::GET         ]],
-	'/statusnet/config[.{extension:json|xml|rss|atom}]'                => [Module\Api\GNUSocial\GNUSocial\Config::class,  [R::GET         ]],
-	'/statusnet/conversation[.{extension:json|xml|rss|atom}]'          => [Module\Api\Friendica\Index::class,             [R::GET         ]],
-	'/statusnet/conversation/{id:\d+}[.{extension:json|xml|rss|atom}]' => [Module\Api\Friendica\Index::class,             [R::GET         ]],
-	'/statusnet/version[.{extension:json|xml|rss|atom}]'               => [Module\Api\GNUSocial\GNUSocial\Version::class, [R::GET         ]],
+	'/media/upload[.{extension:json|xml|rss|atom}]'                    => [Module\Api\Friendica\Index::class,                  [        R::POST]],
+	'/media/metadata/create[.{extension:json|xml|rss|atom}]'           => [Module\Api\Friendica\Index::class,                  [        R::POST]],
+	'/saved_searches/list[.{extension:json|xml|rss|atom}]'             => [Module\Api\Twitter\SavedSearches::class,            [R::GET         ]],
+	'/search/tweets[.{extension:json|xml|rss|atom}]'                   => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
+	'/search[.{extension:json|xml|rss|atom}]'                          => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
+	'/statusnet/config[.{extension:json|xml|rss|atom}]'                => [Module\Api\GNUSocial\GNUSocial\Config::class,       [R::GET         ]],
+	'/statusnet/conversation[.{extension:json|xml|rss|atom}]'          => [Module\Api\Friendica\Index::class,                  [R::GET         ]],
+	'/statusnet/conversation/{id:\d+}[.{extension:json|xml|rss|atom}]' => [Module\Api\GNUSocial\Statusnet\Conversation::class, [R::GET         ]],
+	'/statusnet/version[.{extension:json|xml|rss|atom}]'               => [Module\Api\GNUSocial\GNUSocial\Version::class,      [R::GET         ]],
 
 	'/statuses' => [
 		'/destroy[.{extension:json|xml|rss|atom}]'                 => [Module\Api\Friendica\Index::class,                        [R::DELETE, R::POST]],
@@ -137,10 +137,10 @@ $apiRoutes = [
 	],
 
 	'/users' => [
-		'/lookup[.{extension:json|xml|rss|atom}]'                  => [Module\Api\Friendica\Index::class,    [R::GET         ]],
-		'/search[.{extension:json|xml|rss|atom}]'                  => [Module\Api\Friendica\Index::class,    [R::GET         ]],
-		'/show[.{extension:json|xml|rss|atom}]'                    => [Module\Api\Twitter\Users\Show::class, [R::GET         ]],
-		'/show/{id:\d+}[.{extension:json|xml|rss|atom}]'           => [Module\Api\Friendica\Index::class,    [R::GET         ]],
+		'/lookup[.{extension:json|xml|rss|atom}]'                  => [Module\Api\Twitter\Users\Lookup::class, [R::GET         ]],
+		'/search[.{extension:json|xml|rss|atom}]'                  => [Module\Api\Twitter\Users\Search::class, [R::GET         ]],
+		'/show[.{extension:json|xml|rss|atom}]'                    => [Module\Api\Twitter\Users\Show::class,   [R::GET         ]],
+		'/show/{id:\d+}[.{extension:json|xml|rss|atom}]'           => [Module\Api\Twitter\Users\Show::class,   [R::GET         ]],
 	],
 ];
